2022/vital/pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "allainclair", email = "allainclair@gmail.com"},
]
dependencies = [
    "requests>=2.28.1",
    "fastapi>=0.79.0",
    "uvicorn>=0.18.2",
    "python-dotenv>=0.20.0",
    "pyhumps>=3.7.2",
]
requires-python = ">=3.10"
license = {text = "MIT"}
[project.optional-dependencies]

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
[tool.pdm.dev-dependencies]
test = [
    "pytest>=7.1.2",
    "icecream>=2.1.3",
    "ipython>=8.4.0",
]

[tool.pdm.scripts]
start = "uvicorn api:app --reload"
